--- src/app/_document.tsx
@@ -1,16 +1,0 @@
-// import { Html, Head, Main, NextScript } from "next/document";
-// import Script from "next/script";
-
-// export default function Document() {
-//   const KakaoMapSrc = `//dapi.kakao.com/v2/maps/sdk.js?appkey=${process.env.NEXT_PUBLIC_KAKAOMAP_KEY}&libraries=services,clusterer&autoload=false`;
-//   return (
-//     <Html>
-//       <Head />
-//       <body>
-//         <Main />
-//         <NextScript />
-//         <Script src={KakaoMapSrc} strategy="beforeInteractive" />
-//       </body>
-//     </Html>
-//   );
-// }
