+++ src/app/layout.tsx
@@ -3,11 +3,18 @@ import Header from "@/components/Header";
 import supabase from "@/libs/supabase";
 import "@/styles/globals.css";
 import type { Metadata } from "next";
+import Script from "next/script";
 
 export const metadata: Metadata = {
   title: "Savers",
   description: "Generated by create next app",
 };
+
+declare global {
+  interface Window {
+    kakao: any;
+  }
+}
 
 export default function RootLayout({
   children,
@@ -22,6 +29,13 @@ export default function RootLayout({
           {children}
         </QueryProvider>
       </body>
+      {/* 해당 부분에 Script를 추가한다. */}
+      <Script src="https://developers.kakao.com/sdk/js/kakao.js" async />
+      {/* 이 두번째 Script는 kakao map을 이용하기 위한 Script이다. appkey 부분엔 발급받은 본인의 API KEY를 입력한다. */}
+      <Script
+        type="text/javascript"
+        src={`//dapi.kakao.com/v2/maps/sdk.js?appkey=${process.env.NEXT_PUBLIC_KAKAOMAP_KEY}&autoload=false&libraries=services`}
+      />
     </html>
   );
 }
